requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
audioread==3.0.1
autobahn==24.4.2
Automat==24.8.1
av==14.2.0
babel==2.17.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
csvw==3.5.1
Cython==3.0.12
daphne==4.1.2
decorator==5.2.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.6
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dlinfo==2.0.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-texttospeech==2.25.0
googleapis-common-protos==1.69.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isodate==0.7.2
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langdetect==1.0.9
language-tags==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
msgpack==1.1.0
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
openai==1.65.2
packaging==24.2
phonemizer==3.3.0
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
rdflib==7.1.3
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.23.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
segments==2.3.0
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
threadpoolctl==3.5.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn
wheel==0.45.1
whitenoise==6.9.0
zope.interface==7.2
